+++ src/ecc/curve.rs
@@ -44,7 +44,7 @@ mod tests {
 
     use num::BigUint;
 
-    use crate::elliptic_curve::finite_field::FiniteField;
+    use crate::ecc::finite_field::FiniteField;
 
     use super::*;
