+++ src/lib.rs
@@ -1,4 +1,4 @@
-pub mod elliptic_curve;
+pub mod ecc;
 pub mod script;
 pub mod transactions;
 pub mod utils;
